Add my results of the exercices
This is the results of my work for the udacity course 120
Intro into Data Science.

I forgot to commit after each atomic step (and to clean my code), so it
's a last chance commit.

This is a no go in the industry. May the recruiters fogive me!

(I was simply not thinking to communicate this work when I did it)